% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cardio}
\alias{cardio}
\title{Cardiotocography}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Cardiotocography}
}
\description{
Cardiotocography data from UCI machine learning repository. Raw data have been 
cleaned and an \code{outcome} column added that is a binary variable of predicting
NSP (described below) = 2.
}
\details{
Data Set Information:
2126 fetal cardiotocograms (CTGs) were automatically processed and the respective diagnostic features measured. 
The CTGs were also classified by three expert obstetricians and a consensus classification label assigned to each of them. Classification was both with respect to a morphologic pattern (A, B, C. ...) and to a fetal state (N, S, P). Therefore the dataset can be used either for 10-class or 3-class experiments.

Attribute Information:

LB - FHR baseline (beats per minute) 

AC - # of accelerations per second 

FM - # of fetal movements per second 

UC - # of uterine contractions per second 

DL - # of light decelerations per second 

DS - # of severe decelerations per second 

DP - # of prolongued decelerations per second 

ASTV - percentage of time with abnormal short term variability 

MSTV - mean value of short term variability 

ALTV - percentage of time with abnormal long term variability 

MLTV - mean value of long term variability 

Width - width of FHR histogram 

Min - minimum of FHR histogram 

Max - Maximum of FHR histogram 

Nmax - # of histogram peaks 

Nzeros - # of histogram zeros 

Mode - histogram mode 

Mean - histogram mean 

Median - histogram median 

Variance - histogram variance 

Tendency - histogram tendency 

CLASS - FHR pattern class code (1 to 10) 

NSP - fetal state class code (N=normal; S=suspect; P=pathologic)
}
\references{
Ayres de Campos et al. (2000) SisPorto 2.0 A Program for Automated Analysis of Cardiotocograms. J Matern Fetal Med 5:311-318
}
\keyword{data}
