% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{dbarts_wrapper}
\alias{dbarts_wrapper}
\title{Wrapper for fitting Bayesian additive regression trees (BART) via \code{dbarts}.}
\usage{
dbarts_wrapper(test, train, sigest = NA, sigdf = 3, sigquant = 0.9,
  k = 2, power = 2, base = 0.95, binaryOffset = 0, ntree = 200,
  ndpost = 1000, nskip = 100, printevery = 100, keepevery = 1,
  keeptrainfits = TRUE, usequants = FALSE, numcut = 100,
  printcutoffs = 0, nthread = 1, keepcall = TRUE, verbose = FALSE)
}
\arguments{
\item{test}{A list with named objects \code{Y} and \code{X} (see description).}

\item{train}{A list with named objects \code{Y} and \code{X} (see description).}

\item{sigest}{See \link[dbarts]{dbarts}}

\item{sigdf}{See \link[dbarts]{dbarts}}

\item{sigquant}{See \link[dbarts]{dbarts}}

\item{k}{See \link[dbarts]{dbarts}}

\item{power}{See \link[dbarts]{dbarts}}

\item{base}{See \link[dbarts]{dbarts}}

\item{binaryOffset}{See \link[dbarts]{dbarts}}

\item{ntree}{See \link[dbarts]{dbarts}}

\item{ndpost}{See \link[dbarts]{dbarts}}

\item{nskip}{See \link[dbarts]{dbarts}}

\item{printevery}{See \link[dbarts]{dbarts}}

\item{keepevery}{See \link[dbarts]{dbarts}}

\item{keeptrainfits}{See \link[dbarts]{dbarts}}

\item{usequants}{See \link[dbarts]{dbarts}}

\item{numcut}{See \link[dbarts]{dbarts}}

\item{printcutoffs}{See \link[dbarts]{dbarts}}

\item{nthread}{See \link[dbarts]{dbarts}}

\item{keepcall}{See \link[dbarts]{dbarts}}

\item{verbose}{See \link[dbarts]{dbarts}}
}
\value{
A list with named objects (see description).
}
\description{
Compatible learner wrappers for this package should have a specific format.
Namely they should take as input a list called \code{train} that contains
named objects \code{$Y} and \code{$X}, that contain, respectively, the outcomes
and predictors in a particular training fold. Other options may be passed in
to the function as well. The function must output a list with the following
named objects: \code{test_pred} = predictions of \code{test$Y} based on the learner
fit using \code{train$X}; \code{train_pred} = prediction of \code{train$Y} based 
on the learner fit using \code{train$X}; \code{model} = the fitted model (only 
necessary if you desire to look at this model later, not used for internal 
computations); \code{train_y} = a copy of \code{train$Y}; \code{test_y} = a copy
of \code{test$Y}.
}
\details{
This particular wrapper implements Bayesian additive regression trees using 
\link[dbarts]{dbarts}. We refer readers to the original package's 
documentation for more details.
}
\examples{
# simulate data
Q0 <- function(x){ plogis(x) }
# make list of training data
train_X <- data.frame(x1 = runif(50))
train_Y <- rbinom(50, 1, Q0(train_X$x1))
train <- list(Y = train_Y, X = train_X)
# make list of test data
test_X <- data.frame(x1 = runif(50))
test_Y <- rbinom(50, 1, Q0(train_X$x1))
test <- list(Y = test_Y, X = test_X)
# fit dbarts
dbarts_wrap <- dbarts_wrapper(train = train, test = test)
}
