% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{.Dy}
\alias{.Dy}
\title{Compute one of the terms of the efficient influence function}
\usage{
.Dy(full_long_data, y)
}
\arguments{
\item{full_long_data}{A long form data set}

\item{y}{Which portion of the EIF to compute}
}
\value{
Vector of one piece of EIF evaluated at estimates in \code{full_long_data}
}
\description{
Compute one of the terms of the efficient influence function
}
