% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_functions.R
\name{.get_nested_cv_quantile}
\alias{.get_nested_cv_quantile}
\title{Helper function to get quantile for a single training fold data 
when nested CV is used.}
\usage{
.get_nested_cv_quantile(x, p, prediction_list, folds, quantile_type = 8)
}
\arguments{
\item{x}{An entry in prediction_list.}

\item{p}{The quantile to get.}

\item{prediction_list}{Properly formatted list of predictions.}

\item{folds}{Cross-validation fold assignments.}

\item{quantile_type}{The type of quantile estimate to use.}
}
\description{
Helper function to get quantile for a single training fold data 
when nested CV is used.
}
