% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_functions.R
\name{.get_one_fold}
\alias{.get_one_fold}
\title{Helper function to get results for a single cross-validation fold}
\usage{
.get_one_fold(x, sens, gn, quantile_type = 8, ...)
}
\arguments{
\item{x}{An entry in prediction_list.}

\item{sens}{The sensitivity constraint.}

\item{gn}{An estimate of the marginal probability that \code{Y = 1}.}

\item{quantile_type}{The type of quantile estimate to use.}

\item{...}{Other options (not currently used)}
}
\description{
Helper function to get results for a single cross-validation fold
}
