% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{.get_psi_distribution_nested_cv}
\alias{.get_psi_distribution_nested_cv}
\title{Compute the conditional (given Y = y) CV-estimated distribution of psi}
\usage{
.get_psi_distribution_nested_cv(x, y, prediction_list, folds, epsilon = 0)
}
\arguments{
\item{x}{The outer validation fold withheld}

\item{y}{What value of Y to compute dist. est.}

\item{prediction_list}{List output from .get_predictions.}

\item{folds}{Cross validation fold indicator.}

\item{epsilon}{A vector of estimated coefficients form tmle fluctuation 
submodels.}
}
\value{
A data.frame with the distribution of psi given Y = y with names
psix (what value estimates are evaluated at), dFn (density estimates),
Fn (cdf estimates)
}
\description{
Compute the conditional (given Y = y) CV-estimated distribution of psi
}
