% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_functions.R
\name{.get_quantile}
\alias{.get_quantile}
\title{Helper function to get quantile for a single training fold data 
when nested CV is NOT used.}
\usage{
.get_quantile(x, p, quantile_type = 8)
}
\arguments{
\item{x}{An entry in prediction_list.}

\item{p}{The quantile to get.}

\item{quantile_type}{The type of quantile estimate to use.}
}
\description{
Helper function to get quantile for a single training fold data 
when nested CV is NOT used.
}
