% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_functions.R
\name{.getCVEstimator}
\alias{.getCVEstimator}
\title{Helper function to turn prediction_list into CV estimate of SCRNP}
\usage{
.getCVEstimator(prediction_list, sens, gn, quantile_type = 8, ...)
}
\arguments{
\item{prediction_list}{Properly formatted list of predictions.}

\item{sens}{The sensitivity constraint.}

\item{gn}{The marginal probability that \code{Y = 1}.}

\item{quantile_type}{The type of quantile estimate to use.}

\item{...}{Other options (not currently used)}
}
\description{
Helper function to turn prediction_list into CV estimate of SCRNP
}
