% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_functions.R
\name{.make_targeting_data}
\alias{.make_targeting_data}
\title{Helper function for making data set in proper format for CVTMLE}
\usage{
.make_targeting_data(
  x,
  prediction_list,
  quantile_list,
  density_list,
  folds,
  nested_cv = FALSE,
  gn
)
}
\arguments{
\item{x}{A numeric identifier of which entry in \code{prediction_list} to operate on.}

\item{prediction_list}{Properly formatted list of predictions.}

\item{quantile_list}{List of estimated quantile for each fold.}

\item{density_list}{List of density estimates for each fold.}

\item{folds}{Cross-validation fold assignments.}

\item{nested_cv}{A boolean indicating whether nested CV was used in estimation.}

\item{gn}{An estimate of the marginal probability that \code{Y = 1}.}
}
\description{
Helper function for making data set in proper format for CVTMLE
}
