% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{F_nBn_star}
\alias{F_nBn_star}
\title{Compute the targeted conditional cumulative distribution of the learner at a point}
\usage{
F_nBn_star(psi_x, y, train_pred, train_y, epsilon = 0, tol = 0.001)
}
\arguments{
\item{psi_x}{Value to compute conditional (on Y=y) cdf of learner}

\item{y}{Value of Y to condition on}

\item{train_pred}{Values of Psi_nBn(X) from training sample}

\item{train_y}{Values of Y from training sample}

\item{epsilon}{Vector of fluctuation parameter estimates}

\item{tol}{Truncation level for logistic transformation}
}
\value{
Numeric value of CDF at \code{psi_x}
}
\description{
Compute the targeted conditional cumulative distribution of the learner at a point
}
