% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{F_nBn_star_nested_cv}
\alias{F_nBn_star_nested_cv}
\title{Compute the targeted conditional cumulative distribution of the learner at a point
where the initial distribution is based on cross validation}
\usage{
F_nBn_star_nested_cv(
  psi_x,
  y,
  inner_valid_prediction_and_y_list,
  epsilon = 0,
  tol = 0.001
)
}
\arguments{
\item{psi_x}{Value to compute conditional (on Y=y) cdf of learner}

\item{y}{Value of Y to condition on}

\item{inner_valid_prediction_and_y_list}{A list of predictions and y's from \code{.get_predictions}.}

\item{epsilon}{Vector of fluctuation parameter estimates}

\item{tol}{A truncation level when taking logit transformations.}
}
\value{
Numeric value of CDF at \code{psi_x}
}
\description{
Compute the targeted conditional cumulative distribution of the learner at a point
where the initial distribution is based on cross validation
}
