% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{fluc_mod_optim_1}
\alias{fluc_mod_optim_1}
\title{Helper function for CVTMLE grid search}
\usage{
fluc_mod_optim_1(epsilon, fld, tol = 0.001)
}
\arguments{
\item{epsilon}{Fluctuation parameter}

\item{fld}{full_long_data_list}

\item{tol}{Tolerance on predictions close to 0 or 1}
}
\value{
A numeric value of negative log-likelihood
}
\description{
Helper function for CVTMLE grid search
}
