% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scrnp}
\alias{print.scrnp}
\title{Print results of cv_scrnp}
\usage{
\method{print}{scrnp}(x, se_type = "std", ci_level = 0.95, ...)
}
\arguments{
\item{x}{An object of class "cvauc"}

\item{se_type}{The type of standard error (currently only "std")}

\item{ci_level}{Level of confidence interval to print. Defaults to 0.95.}

\item{...}{Other options (not currently used)}
}
\description{
Print results of cv_scrnp
}
